Add docstrings for fixtures and move user workflow fixtures into separate files
