+++ tests/profiler/conftest.py
@@ -14,10 +14,10 @@ from great_expectations.profiler.parameter_builder.parameter_container import (
     ParameterNode,
 )
 from great_expectations.profiler.rule.rule import Rule
-from tests.profiler.bob_user_workflow_fixture import bob_columnar_table_multi_batch
 
 # noinspection PyUnresolvedReferences
-from tests.profiler.user_workflow_fixtures import alice_columnar_table_single_batch
+from tests.profiler.alice_user_workflow_fixture import alice_columnar_table_single_batch
+from tests.profiler.bob_user_workflow_fixture import bob_columnar_table_multi_batch
 
 
 # noinspection PyPep8Naming
